requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-cors-headers==3.1.1
django-oauth-toolkit==1.2.0
djangorestframework==3.10.3
idna==2.8
oauthlib==3.1.0
psycopg2==2.8.4
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
